+++ gba_test_macros/tests/trybuild/result_should_panic.rs
@@ -1,0 +1,12 @@
+#![feature(custom_test_frameworks)]
+
+use gba_test_macros::test;
+
+#[test]
+#[should_panic]
+fn foo() -> Result<(), ()> {
+    Ok(())
+}
+
+
+fn main() {}